Internal Function

IBD Summary Table

getConditionSNP(genotypes, snp.ld, maximum.ld.r2)

Arguments

genotypes
A data frame containing genotype information for a set of SNPs. See getGenotypes for more details.

snp.ld
A data frame generated from PLINK containing information on LD between pairs of SNPs (A and B). This data frame contains exactly 7 columns of informations:

Chromosome of SNP A (type "numeric" or "integer")

Base-pair position of SNP A (type "numeric" or "integer")

SNP A identifier (type "character")

Chromosome of SNP B (type "numeric" or "integer")

Base-pair position of SNP B (type "numeric" or "integer")

SNP B identifier (type "character")

R-sequared LD statistic between SNP A and SNP B (type "numeric")

where each row contains the LD information for a single pair of SNPs. The data frame should contain the header colnames(snp.ld)=c(CHR_A, BP_A, SNP_A, CHR_B, BP_B, SNP_B, R2)

maximum.ld.r2
A numeric value denoting the maximum linkage disequilibrium R2 value allowed between pairs of SNPs.

Value

A data frame containing columns

Chromosome (type numeric or integer)

SNP identifier (type character)

Genetic map position (type numeric or integer)

Base-pair position (type numeric or integer)

Numeric SNP identifier (type numeric or integer)

Numeric condition SNP identifier (type numeric or integer)

with columns headed chr, snp_id, pos_M, pos_bp, marker.id and cond.snp.id.

Details

Extracts the SNP with the largest LD with the current SNP and returns this as the condition SNP