XIBD Parameter Estimation

Estimate the number of meioses and the probabilities of sharing 0, 1 and 2 alleles IBD between pairs.

getIBDparameters(ped.genotypes, number.cores = 1)

Arguments

ped.genotypes
a named list containing pedigree, genotypes and model. See Value description in getGenotypes for more details. The family IDs and individual IDs in pedigree must match the family IDs and individual IDs in the header of genotypes.

number.cores
the number of cores used for parallel execution.

Value

A named list containing autosome_parameters and X_chromosome_parameters. Both autosome_parameters and X_chromosome_parameters are data frames with columns:

Family 1 ID (type "character")

Individual 1 ID (type "character")

Family 2 ID (type "character")

Individual 2 ID (type "character")

The number of meiosis (type "numeric" or "integer")

Probability of sharing 0 alleles IBD (type "numeric")

Probability of sharing 1 allele IBD (type "numeric")

Probability of sharing 2 alleles IBD (type "numeric")

headed fid1, iid1, fid2, iid2, m, ibd0, ibd1 and ibd2, respectively. Each row describes parameters for a unique pair of samples. autosome_parameters will be NULL if chromosomes 1-22 are all excluded from ped.genotypes and X_chromosome_parameters will be NULL if chromosome 23 is excluded from ped.genotypes.

Examples

# estimate parameters
my_parameters <- getIBDparameters(ped.genotypes = example_genotypes,
                                  number.cores = 1)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   2%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%

str(my_parameters)

#> List of 2
#>  $ autosome_parameters    :'data.frame':	45 obs. of  8 variables:
#>   ..$ fid1: chr [1:45] "1" "1" "1" "1" ...
#>   ..$ iid1: chr [1:45] "1" "1" "1" "1" ...
#>   ..$ fid2: chr [1:45] "1" "1" "1" "1" ...
#>   ..$ iid2: chr [1:45] "2" "4" "5" "8" ...
#>   ..$ m   : num [1:45] 6.06 8.59 1 7.98 10.04 ...
#>   ..$ ibd0: num [1:45] 0.97 0 0 0.665 0.65 ...
#>   ..$ ibd1: num [1:45] 0.0299 0.9896 1 0.3188 0.3464 ...
#>   ..$ ibd2: num [1:45] 0 0.0104 0 0.0158 0.0038 0.0021 0 0.0013 0 0.0079 ...
#>  $ X_chromosome_parameters:'data.frame':	45 obs. of  8 variables:
#>   ..$ fid1: chr [1:45] "1" "1" "1" "1" ...
#>   ..$ iid1: chr [1:45] "1" "1" "1" "1" ...
#>   ..$ fid2: chr [1:45] "1" "1" "1" "1" ...
#>   ..$ iid2: chr [1:45] "2" "4" "5" "8" ...
#>   ..$ m   : num [1:45] 6.41 5.68 1 1 1.9 ...
#>   ..$ ibd0: num [1:45] 0.977 0.961 0 0 0.463 ...
#>   ..$ ibd1: num [1:45] 0.0235 0.0389 1 1 0.5374 ...
#>   ..$ ibd2: num [1:45] 0 0 0 0 0 0 0 0 0 0 ...