XIBD Proportion of Pairs IBD

Calculate the proportion of pairs IBD at each SNP.

getLocusProportion(ped.genotypes, locus.matrix, groups = NULL)

Arguments

ped.genotypes
a named list containing pedigree, genotypes and model. See Value description in getGenotypes for more details. The family IDs and individual IDs in pedigree must match the family IDs and individual IDs in the header of genotypes.

locus.matrix
a data frame containing the binary IBD information for each SNP and each pair. See value description in getLocusMatrix for more details.

groups
a data frame with three columns of information:

Family ID

Sample ID

Group ID

Group ID, for example, can be affection status. If groups is specified, each sample in the pedigree should belong to a group, and locus proportions will be calculated for each group. The default is groups=NULL.

Value

A data frame the following 5 columns:

Chromosome (type "character", "numeric" or "integer")

SNP identifiers (type "character")

Genetic map distance (centi morgans, cM) (type "numeric")

Base-pair position (type "integer")

Proportion of pairs IBD (type "integer")

where each row describes a unique SNP. The data frame is headed chr, snp_id, pos_M, pos_bp and prop respectively. If groups is specified then there will be one column of proportions of each combination of groups with these column header by group IDs and the number of pairs in the group.

Examples

# generate a binary IBD matrix
my_locus_matrix <- getLocusMatrix(ped.genotypes = example_genotypes,
                                  ibd.segments = example_ibd)

# calculate the proportion of pairs IBD at each SNP
my_locus_prop <- getLocusProportion(ped.genotypes = example_genotypes,
                                    locus.matrix = my_locus_matrix,
                                    groups = NULL)